Add defensive code in case mechanism elements are mistakenly chain elmements
mechanisms.NewServer(map[string]networkservice.NetworkServiceServer) presumes
that the provided implementations for a mechanism *don't* chain, but simply return to it when they are done.

The probability of an implementor getting that right is basically zero, so we wrap them in a chain
to make sure that if they call next.Server(ctx).{Request/Close} they don't continue down the
chain that mechanisms.NewServer is on.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>