Add injectpeer NetworkServiceClient chain element
It  is desirable to know the grpc.Peer information of your server, so for example, you can add
a proper 'aud' claim to a JWT Token.
golang grpc doesn't provide the grpc.Peer information to clients unless
(1)  They ask for it with the grpc.Peer grpc.CallOption
(2)  Even then, only on return.
This chain element compensates for that by:
(1)  Adding the grpc.Peer grpc.CallOption to its call to next.Client(ctx)
(2)  Keeping a running record on the return of the grpc.Peer
(3)  Injecting that grpc.Peer into the context used to call next.Client(ctx) once it has a grpc.Peer on record
(4)  If and only if it it doesn't havea grpc.Peer and its call to next.Client(ctx) return an error, tries doing
     the call again *with* the grpc.Peer added to the context to see if that clears up the error.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>