Fix connect chain element to correctly chain together connections.
Connect fundamentally allows a server to connect to a client.

This PR fixes two issues:

1)  The Connection.Id we index by must be the Request.Connection arriving
at the Server, not the Connection.Id returned from the client.  This is
necessary to be able to close correctly.
2)  Connect needs to call next.Client(ctx) and get a response before it
moves on to next.Server(ctx).  This was being done in Request, but not
in Close.  Because it wasn't being done in Close, teardown could not be done
properly for xconnect.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>