Rework heal chain element to support updatetoken restoring Connection.Id to the Path.Index
The updatetoken chain element should restore the Connection.Id to match Path.PathSegment[Index].

Additionally, updatetoken should be the last thing that happens before
sending out the RPC call, so as to insure the Connection.Id and Path.Index
are always for the Connection.Id within the client when in client chain
elements.

This is important because otherwise the Connect.Id on the return through the
client does not match the Request.Connection.Id, and various indexing by identifiers
fails to work properly (including things like heal etc).

Because injectpeer is simply present to enable updatetoken, it is also moved
to the end of the chain just before updatetoken.

Unfortunately, this revealed a huge number of race conditions in the heal chain element,
which has been completely reworked to dispel them.

In doing so, it was discovered that eventchannel//monitor_client.go and monitor_connection_client.go
has some issues around respecting context.Context.

Which resulted in some small changes to adapters/monitor_server_to_client.go

Signed-off-by: Ed Warnicke <hagbard@gmail.com>