Rework client, connect, to simplify and prepare for simpler healing.
Previously we had a single 'client' per grpc.ClientConnInterface.
This created a lot of complexity, particularly in connect.

This rework of client moves to a single 'client' that can handle
multiple grpc.ClientConnInterfaces.

This vastly simplified both the client chain and connect chain.

Finally, heal has been temporarily removed in preparation for a rework
of heal focused on back propogating Connection 'DOWN' events via
monitor.  The net net at the end of the process should be:

1)  Refresh is only originated by actual clients, not the 'client' part
of passthrough NSEs
2)  Heal will only be originated by actual clients, not the 'client' part
pf passthrough NSEs.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>