Fix backward compatibility for NewNetworkServiceEndpointRegistryClient and NewNetworkServiceRegistryClient
NewNetworkServiceEndpointRegistryClient and NewNetworkServiceRegistryClient where recently enhanced.
In that process, the leading "connectTo *url.URL" parameter was removed.
The functionality continues to be supported via the WithClientURL option.

In order to avoid breaking current downstream users, the leading "connectTo *url.URL"
has been restored in such a way that it has no effect if it is nil and otherwise
utilize the WithClient option (to insure the code path is exercised).

Signed-off-by: Ed Warnicke <hagbard@gmail.com>