Don't trace gRPC "Canceled" errors (#1353)
We were getting a *lot* of log chaff at TRACE level during normal
operation because gRPC returns a "context canceled" error periodically
but it's not really an error for us because we immediately loop around
again.

This commit uses the gRPC status.Code() method to identify these
non-errors. It also uses errors.Is() instead of == to compare error
objects. This handles wrapped errors.

https://pkg.go.dev/google.golang.org/grpc/status#Code

https://github.com/golang/go/wiki/ErrorValueFAQ#how-should-i-change-my-error-handling-code-to-work-with-the-new-features

Signed-off-by: Toby Cabot <toby@acnodal.io>

Signed-off-by: Toby Cabot <toby@acnodal.io>
Co-authored-by: toby cabot <toby@acnodal.io>