Enhance FakeDNSResolver to track IP addresses
FakeDNSResolver used to track only ports, addresses were always
127.0.0.1. I'd like to write some unit tests for IPV6 so I need
FakeDNSResolver to store and retrieve any address.

As part of this work now fakeDNSResolver is better hidden, and no
longer exported: clients/tests use NewFakeResolver() as a
constructor, which returns a Resolver, and call AddSRVEntry() to add
entries. This allows all client/test code to use Resolver exclusively.

common_test.go appeared to be dead code so I deleted it. It had
another implementation of a fake Resolver but no tests.

Signed-off-by: Toby Cabot <toby@acnodal.io>